refactor: extract date-filters module for lib-adapter
